Update bindgen to 0.71
This commit updates bindgen to 0.71. It also adds a test that verifies that the bundled bindings match the generated ones. Additionally it cleans up the generated bindings to only include items that are relevant for libpq
