+++ Cargo.toml
@@ -6,6 +6,8 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/sgrif/pq-sys"
 links = "pq"
 build = "build.rs"
+edition = "2021"
+rust-version = "1.82.0"
 
 [workspace]
 members = ["pq-src"]
@@ -15,13 +17,18 @@ name = "pq_sys"
 
 [dependencies]
 pq-src = { path = "pq-src", version = ">=0.2, <0.4", optional = true , default-features = false }
+libc = "0.2.100"
 
 [build-dependencies]
 pkg-config = { version = "0.3.0", optional = true }
-bindgen = { version = "0.69.1", optional = true }
+bindgen = { version = "0.71.0", optional = true }
 
 [target.'cfg(target_env = "msvc")'.build-dependencies]
 vcpkg = "0.2.6"
+
+[dev-dependencies]
+similar-asserts = "1.6.0"
+bindgen = "0.71.0"
 
 [features]
 default = []
