+++ src/make_bindings.rs
@@ -1,0 +1,20 @@
+/* keep in sync with the command in DEVELOPMENT.md */
+bindgen::Builder::default()
+    .rustified_enum(".*")
+    .header("wrapper.h")
+    .allowlist_var("PG_.*")
+    .allowlist_var("LIBPQ_.*")
+    .allowlist_var("PQ.*")
+    .allowlist_type("Oid")
+    .allowlist_type("ConnStatusType")
+    .allowlist_type("Postgres.*")
+    .allowlist_type("pg.*")
+    .allowlist_type("PG.*")
+    .allowlist_type("PQ.*")
+    .allowlist_type("pq.*")
+    .allowlist_function("PQ.*")
+    .allowlist_function("lo_.*")
+    .allowlist_function("pg_.*")
+    .opaque_type("FILE")
+    .blocklist_type("FILE")
+    .raw_line("use libc::FILE;")
